+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added support for ingesting the following **new** relationships:
+
+  | Source     | \_class     | Target                       |
+  | ---------- | ----------- | ---------------------------- |
+  | `azure_vm` | `GENERATED` | `azure_shared_image_version` |
+
 ## 5.29.0 - 2021-06-07
 
 ### Added